requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.2.1
django-debug-toolbar-template-timings==0.6.4
ipdb==0.8
django-extensions==1.3.11
flake8==2.2.2
mkdocs==0.9
